+++ Magent/App/SplitViewController.swift
@@ -223,7 +223,7 @@ final class SplitViewController: NSSplitViewController {
             contentContainerVC.setContent(detailVC)
         }
 
-        if AppFeatures.jiraIntegrationEnabled, thread.jiraUnassigned {
+        if AppFeatures.jiraSyncEnabled, thread.jiraUnassigned {
             BannerManager.shared.show(
                 message: "This ticket is no longer assigned to you",
                 style: .info,